Changed access types of classes that were internal to public.
Created a UserRecord class which facilitates the use of an SQlite database. Not sure how it would work once it is on a different device but that needs to be tested.

Created a UserRepository class which allows for the manipulation of the User class using the methods present in the class.

Removed the specific IDs for the User subclasses to simplify the application.